[7] Restructured main file
- Moved image processing operations
to operations.rs
- Cleanup of main file

%SOFTWARE
